looking at bimodal generation

+++ bimodalGenerationTime.c
@@ -11,17 +11,17 @@ int main(int argc, char ** argv){
 	int rank, numProc;
 	double start, end, time;
 
-	unsigned long long int arraySize = 100000000;
+	  int arraySize = 100000000;
 
         MPI_Init(&argc, &argv);
         MPI_Comm_rank(MPI_COMM_WORLD, &rank);
         MPI_Comm_size(MPI_COMM_WORLD, &numProc);
 	
 
-	unsigned long long int* integers = (unsigned long long int*) malloc(arraySize * 8);
+	  int * integers = (  int *) malloc(arraySize * 8);
 
 	start = MPI_Wtime();// Start Timing
-        unsigned long long int i;
+          int  i;
         for( i = 0; i < arraySize; i++){
                 integers[i] = i;
         }
